# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ClubHub API — multi-team soccer club management
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /teams:
    get:
      operationId: listTeams
      tags: [teams]
      summary: List teams visible to the current user
      responses:
        "200":
          description: Teams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Team"
    post:
      operationId: createTeam
      tags: [teams]
      summary: Create a team
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
  /teams/{teamId}:
    get:
      operationId: getTeam
      tags: [teams]
      summary: Get a team
      parameters:
        - $ref: "#/components/parameters/TeamId"
      responses:
        "200":
          description: Team
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateTeam
      tags: [teams]
      summary: Update a team
      parameters:
        - $ref: "#/components/parameters/TeamId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
        "404":
          $ref: "#/components/responses/NotFound"
  /teams/{teamId}/summary:
    get:
      operationId: getTeamSummary
      tags: [teams]
      summary: Team dashboard summary (next event, availability, recent activity)
      parameters:
        - $ref: "#/components/parameters/TeamId"
      responses:
        "200":
          description: Team summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamSummary"
  /teams/{teamId}/members:
    get:
      operationId: listTeamMembers
      tags: [teams]
      summary: List members of a team
      parameters:
        - $ref: "#/components/parameters/TeamId"
      responses:
        "200":
          description: Members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TeamMember"
    post:
      operationId: addTeamMember
      tags: [teams]
      summary: Add a person to a team with a role
      parameters:
        - $ref: "#/components/parameters/TeamId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamMemberInput"
      responses:
        "201":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamMember"
  /team-members/{memberId}:
    patch:
      operationId: updateTeamMember
      tags: [teams]
      summary: Update a team membership (role, number, position)
      parameters:
        - $ref: "#/components/parameters/MemberId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamMemberUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamMember"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: removeTeamMember
      tags: [teams]
      summary: Remove a person from a team
      parameters:
        - $ref: "#/components/parameters/MemberId"
      responses:
        "204":
          description: Removed
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  parameters:
    TeamId:
      name: teamId
      in: path
      required: true
      schema:
        type: integer
    MemberId:
      name: memberId
      in: path
      required: true
      schema:
        type: integer
...
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Person:
      type: object
      required: [id, firstName, lastName, fullName, isMinor]
      properties:
        id: { type: integer }
        firstName: { type: string }
        lastName: { type: string }
        fullName: { type: string }
        email: { type: ["string", "null"] }
        phone: { type: ["string", "null"] }
        avatarUrl: { type: ["string", "null"] }
        dateOfBirth: { type: ["string", "null"] }
        isMinor: { type: boolean }
        hasLogin: { type: boolean }
...
    Team:
      type: object
      required: [id, name, ageGroup, playerCount, staffCount]
      properties:
        id: { type: integer }
        name: { type: string }
        ageGroup: { type: string }
        gender: { type: ["string", "null"] }
        colorHex: { type: ["string", "null"] }
        seasonId: { type: ["integer", "null"] }
        seasonName: { type: ["string", "null"] }
        playerCount: { type: integer }
        staffCount: { type: integer }
    TeamSummary:
      type: object
      required: [team, playerCount, goingCount, maybeCount, outCount, noResponseCount, recentPostCount]
      properties:
        team: { $ref: "#/components/schemas/Team" }
        nextEvent: { $ref: "#/components/schemas/Event" }
        playerCount: { type: integer }
        goingCount: { type: integer }
        maybeCount: { type: integer }
        outCount: { type: integer }
        noResponseCount: { type: integer }
        recentPostCount: { type: integer }
...
    TeamMember:
      type: object
      required: [id, teamId, role, person]
      properties:
        id: { type: integer }
        teamId: { type: integer }
        role: { type: string, enum: [manager, coach, player] }
        jerseyNumber: { type: ["integer", "null"] }
        position: { type: ["string", "null"] }
        person: { $ref: "#/components/schemas/Person" }
...
    Event:
      type: object
      required: [id, teamId, teamName, type, title, startsAt, goingCount, maybeCount, outCount, noResponseCount]
      properties:
        id: { type: integer }
        teamId: { type: integer }
        teamName: { type: string }
        type: { type: string, enum: [training, game, social, other] }
        title: { type: string }
        location: { type: ["string", "null"] }
        opponent: { type: ["string", "null"] }
        startsAt: { type: string }
        endsAt: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        goingCount: { type: integer }
        maybeCount: { type: integer }
        outCount: { type: integer }
        noResponseCount: { type: integer }
        myRsvp: { type: ["string", "null"], enum: [going, maybe, out, null] }
...
    TeamInput:
      type: object
      required: [name, ageGroup]
      properties:
        name: { type: string, minLength: 1 }
        ageGroup: { type: string, minLength: 1 }
        gender: { type: string }
        colorHex: { type: string }
        seasonId: { type: integer }
    TeamUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        ageGroup: { type: string, minLength: 1 }
        gender: { type: string }
        colorHex: { type: string }
        seasonId: { type: integer }
    TeamMemberInput:
      type: object
      required: [personId, role]
      properties:
        personId: { type: integer }
        role: { type: string, enum: [manager, coach, player] }
        jerseyNumber: { type: integer }
        position: { type: string }
    TeamMemberUpdate:
      type: object
      properties:
        role: { type: string, enum: [manager, coach, player] }
        jerseyNumber: { type: integer }
        position: { type: string }
...