# lib/api-zod/src/generated/types/teamUpdate.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * ClubHub API — multi-team soccer club management
 * OpenAPI spec version: 0.1.0
 */

export interface TeamUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  ageGroup?: string;
  gender?: string;
  colorHex?: string;
  seasonId?: number;
}